Moved every function to seperate files
